# docs/openapi/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Proppadia Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/mlb/market-odds": {
      "get": {
        "tags": [
          "mlb"
        ],
        "summary": "MLB market odds lookup (OddsAPI)",
        "operationId": "mlb_market_odds_api_mlb_market_odds_get",
        "parameters": [
          {
            "name": "player_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Player full name",
              "title": "Player Name"
            },
            "description": "Player full name"
          },
          {
            "name": "prop_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "MLB prop type key",
              "title": "Prop Type"
            },
            "description": "MLB prop type key"
          },
          {
            "name": "game_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "YYYY-MM-DD (defaults to today ET)",
              "title": "Game Date"
            },
            "description": "YYYY-MM-DD (defaults to today ET)"
          },
          {
            "name": "over_under",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "over|under",
              "default": "over",
              "title": "Over Under"
            },
            "description": "over|under"
          },
          {
            "name": "line",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Prop line value",
              "title": "Line"
            },
            "description": "Prop line value"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mlb/schedule": {
      "get": {
        "tags": [
          "mlb"
        ],
        "summary": "MLB schedule proxy (backend-owned)",
        "operationId": "mlb_schedule_api_mlb_schedule_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "YYYY-MM-DD (defaults to today ET)",
              "title": "Date"
            },
            "description": "YYYY-MM-DD (defaults to today ET)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/players/resolve": {
      "get": {
        "tags": [
          "mlb"
        ],
        "summary": "Resolve MLB player",
        "operationId": "players_resolve_api_players_resolve_get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Player name",
              "title": "Name"
            },
            "description": "Player name"
          },
          {
            "name": "player_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Alias for name",
              "title": "Player Name"
            },
            "description": "Alias for name"
          },
          {
            "name": "player_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Known player id",
              "title": "Player Id"
            },
            "description": "Known player id"
          },
          {
            "name": "team_abbr",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Team abbreviation",
              "title": "Team Abbr"
            },
            "description": "Team abbreviation"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolvePlayerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/games/context": {
      "get": {
        "tags": [
          "mlb"
        ],
        "summary": "Resolve MLB game context by team/date",
        "operationId": "games_context_api_games_context_get",
        "parameters": [
          {
            "name": "team_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "MLB team id",
              "title": "Team Id"
            },
            "description": "MLB team id"
          },
          {
            "name": "for_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "YYYY-MM-DD (defaults to today ET)",
              "title": "For Date"
            },
            "description": "YYYY-MM-DD (defaults to today ET)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamesContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/players/lookup": {
      "get": {
        "tags": [
          "mlb"
        ],
        "summary": "Lookup MLB player by id",
        "operationId": "players_lookup_api_players_lookup_get",
        "parameters": [
          {
            "name": "player_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "MLB player id",
              "title": "Player Id"
            },
            "description": "MLB player id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolvePlayerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/players/search": {
      "get": {
        "tags": [
          "mlb"
        ],
        "summary": "Search MLB players",
        "operationId": "players_search_api_players_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Name search query",
              "title": "Q"
            },
            "description": "Name search query"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayersSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/players": {
      "get": {
        "tags": [
          "mlb"
        ],
        "summary": "List MLB players",
        "operationId": "players_list_api_players_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 2000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerListItem"
                  },
                  "title": "Response Players List Api Players Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/player-profile/{player_id}": {
      "get": {
        "tags": [
          "mlb"
        ],
        "summary": "MLB player profile",
        "operationId": "get_player_profile_api_player_profile__player_id__get",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prepareProp": {
      "post": {
        "tags": [
          "mlb"
        ],
        "summary": "Prepare MLB prop features",
        "operationId": "prepare_prop_endpoint_api_prepareProp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreparePropRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreparePropResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/predict": {
      "post": {
        "tags": [
          "mlb"
        ],
        "summary": "Predict MLB prop",
        "operationId": "predict_prop_endpoint_api_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/props/add": {
      "post": {
        "tags": [
          "mlb"
        ],
        "summary": "Persist user-added MLB prop from commit token",
        "operationId": "add_prop_endpoint_api_props_add_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPropRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddPropResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GamesContextResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "found": {
            "type": "boolean",
            "title": "Found"
          },
          "team_id": {
            "type": "integer",
            "title": "Team Id"
          },
          "team_abbr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Abbr"
          },
          "for_date": {
            "type": "string",
            "title": "For Date"
          },
          "game_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Game Id"
          },
          "game_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Game Time"
          },
          "is_home": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Home"
          },
          "opponent_team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Team Id"
          },
          "opponent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent"
          },
          "opponent_encoded": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Encoded"
          },
          "game_day_of_week": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Game Day Of Week"
          },
          "time_of_day_bucket": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Of Day Bucket"
          },
          "starting_pitcher_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starting Pitcher Id"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "found",
          "team_id",
          "for_date"
        ],
        "title": "GamesContextResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlayerListItem": {
        "properties": {
          "player_id": {
            "type": "integer",
            "title": "Player Id"
          },
          "player_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player Name"
          },
          "team": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team"
          },
          "last_prop_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Prop Date"
          }
        },
        "type": "object",
        "required": [
          "player_id"
        ],
        "title": "PlayerListItem"
      },
...
      "PlayersSearchResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/PlayerIdentity"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "count",
          "rows"
        ],
        "title": "PlayersSearchResponse"
      },
...
      "ResolvePlayerResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "found": {
            "type": "boolean",
            "title": "Found"
          },
          "player_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player Id"
          },
          "player_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player Name"
          },
          "team_abbr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Abbr"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "matched_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched On"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "found"
        ],
        "title": "ResolvePlayerResponse"
      },
...
    }
  }
}